Fixed the "Select All" issue in Gallery3D
Two issues:
    1) "Select All" doesn't select all videos and photos in one album
    2) "Select All" deselects some items previously selected

The problem is caused by the same IDs in the image table and video table
in database. The ID generation for each item is fine. But we cannot use
ID as the only criteria for the uniqueness of items. If one image has the
id "100" in the image table and one video has the id "100" in the video
table, and both of them happen to be in the same bucket, then only one
will be selected when we do "Select All".

Change-Id: I6431827779a43dfaaf57338627ae3572f1b7a58b

Signed-off-by: madan ankapura <mankapur@sta.samsung.com>